+++ src/main/java/dev/enjarai/blahajtotem/BlahajType.java
@@ -2,8 +2,14 @@ package dev.enjarai.blahajtotem;
 
 import net.minecraft.util.Identifier;
 
-public record BlahajType(String name, Identifier model, int... colors) {
+import java.util.List;
+
+public record BlahajType(String name, List<String> alternatives, Identifier model, int... colors) {
     public BlahajType(String name, int... colors) {
-        this(name, BlahajTotem.id("item/shork"), colors);
+        this(name, List.of(), BlahajTotem.id("item/shork"), colors);
+    }
+
+    public BlahajType(String name, List<String> alternatives, int... colors) {
+        this(name, alternatives, BlahajTotem.id("item/shork"), colors);
     }
 }