+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-- Added `whale` shark variant, turning your shark into our also beloved Blåvingad.
-- Added `prider` shark texture, similar to the `pride` shark, but with just a tad more representation.
-- Tweaked `trans` shark texture to be more in line with other pride textures.
+- Added the option to not have Totems be Blåhaj by default as a builtin resource pack.
+- Fixed a mistake in the updated `trans` shark texture.